+++ src/Delimiter/Processor/StaggeredDelimiterProcessor.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace League\CommonMark\Delimiter\Processor;
 
 use League\CommonMark\Delimiter\DelimiterInterface;
+use League\CommonMark\Exception\InvalidArgumentException;
 use League\CommonMark\Node\Inline\AbstractStringContainer;
 
 /**
@@ -71,7 +72,7 @@ final class StaggeredDelimiterProcessor implements DelimiterProcessorInterface
         $len = $processor->getMinLength();
 
         if (isset($this->processors[$len])) {
-            throw new \InvalidArgumentException(\sprintf('Cannot add two delimiter processors for char "%s" and minimum length %d', $this->delimiterChar, $len));
+            throw new InvalidArgumentException(\sprintf('Cannot add two delimiter processors for char "%s" and minimum length %d', $this->delimiterChar, $len));
         }
 
         $this->processors[$len] = $processor;