+++ src/Extension/FrontMatter/Data/SymfonyYamlFrontMatterParser.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace League\CommonMark\Extension\FrontMatter\Data;
 
+use League\CommonMark\Exception\MissingDependencyException;
 use League\CommonMark\Extension\FrontMatter\Exception\InvalidFrontMatterException;
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Yaml;
@@ -25,7 +26,7 @@ final class SymfonyYamlFrontMatterParser implements FrontMatterDataParserInterfa
     public function parse(string $frontMatter)
     {
         if (! \class_exists(Yaml::class)) {
-            throw new \RuntimeException('Failed to parse yaml: "symfony/yaml" library is missing');
+            throw new MissingDependencyException('Failed to parse yaml: "symfony/yaml" library is missing');
         }
 
         try {