+++ src/Parser/Inline/InlineParserMatch.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace League\CommonMark\Parser\Inline;
 
+use League\CommonMark\Exception\InvalidArgumentException;
+
 final class InlineParserMatch
 {
     private string $regex;
@@ -74,7 +76,7 @@ final class InlineParserMatch
             if ($caseSensitive === null) {
                 $caseSensitive = $definition->caseSensitive;
             } elseif ($caseSensitive !== $definition->caseSensitive) {
-                throw new \LogicException('Case-sensitive and case-insensitive defintions cannot be comined');
+                throw new InvalidArgumentException('Case-sensitive and case-insensitive definitions cannot be combined');
             }
         }